refactored controller
