+++ app/controllers/sessions.rb
@@ -1,0 +1,24 @@
+class MakersBnb < Sinatra::Base
+
+	get '/sessions/new' do
+    erb(:'sessions/new')
+  end
+
+  post '/sessions' do
+    user = User.authenticate(params[:email],
+                             params[:password])
+    if user
+      session[:user_id] = user.id
+      redirect "/space/index"
+    else
+      flash.now[:errors] = ["Incorrect email or password"]
+      erb(:"sessions/new")
+    end
+  end
+
+  delete '/sessions' do
+    session[:user_id] = nil
+    redirect '/home'
+  end
+  
+end
