+++ app/views/space/view.erb
@@ -6,6 +6,6 @@
   <input type="submit" value="Request Booking">
 </form>
 
-<form action="/space/index">
+<!-- <form action="/space/index">
   <button type="submit">Back to Index</button>
-</form>
+</form> -->
